# backend/src/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "description": "Api Documentation",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/product/create": {
            "post": {
                "tags": [
                    "Product"
                ],
                "summary": "Create Product",
                "operationId": "505c8b0302823932c582acbf8e0cff15",
                "requestBody": {
                    "description": "All Request Body For Product",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "indomie"
                                    },
                                    "unite_price": {
                                        "type": "integer",
                                        "example": "3000"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success Created",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid Values",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/product/delete/{code}": {
            "delete": {
                "tags": [
                    "Product"
                ],
                "summary": "Delete Product",
                "operationId": "c88282a286106aa33f5f5c67c6fda119",
                "parameters": [
                    {
                        "name": "code",
                        "in": "path",
                        "description": "Code For Product",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success Deleted",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid Values",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/reports/stock/{month}": {
            "get": {
                "tags": [
                    "Reports"
                ],
                "summary": "Stock Reports",
                "operationId": "3b55bbe8e9d6d31974190a7357239110",
                "parameters": [
                    {
                        "name": "month",
                        "in": "path",
                        "description": "input the param this-month or previous-month",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid Values",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/stock/in/list": {
            "get": {
                "tags": [
                    "Stock"
                ],
                "summary": "Stock in list",
                "operationId": "3f4aee3beb1f4ddf6ccb4cc69d3993ab",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid Values",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/stock/in/update/{id}": {
            "put": {
                "tags": [
                    "Stock"
                ],
                "summary": "Update Stock In",
                "operationId": "cf761ebb5d841a6eb3bdc75511806641",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id For Stock In",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "description": "All Request Body For Stock In",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "product_name": {
                                        "type": "string",
                                        "example": ""
                                    },
                                    "supplier_name": {
                                        "type": "string",
                                        "example": ""
                                    },
                                    "qty": {
                                        "type": "integer",
                                        "example": "0"
                                    },
                                    "date_in": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2024-09-26"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success Updated",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid Values",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "JWT Authorization header using the Bearer scheme.",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}